+++ idf_component_tools/hash_tools/validator.py
@@ -1,12 +1,11 @@
 # SPDX-FileCopyrightText: 2023 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
-import os
 from pathlib import Path
 
 from .calculate import hash_dir
 
 try:
-    from typing import Any, Iterable, Text
+    from typing import Iterable, Text
 except ImportError:
     pass