+++ idf_component_tools/registry/multi_storage_client.py
@@ -1,10 +1,13 @@
-# SPDX-FileCopyrightText: 2022-2023 Espressif Systems (Shanghai) CO LTD
+# SPDX-FileCopyrightText: 2022-2024 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
 """Classes to work with Espressif Component Web Service"""
+from collections import namedtuple
 
 from .api_client_errors import ComponentNotFound, VersionNotFound
 from .component_details import ComponentDetailsWithStorageURL
 from .storage_client import StorageClient
+
+ComponentInfo = namedtuple('ComponentInfo', ['data', 'storage_url'])
 
 
 class MultiStorageClient:
@@ -49,3 +52,20 @@ class MultiStorageClient:
                 error_message = str(err)
 
         raise VersionNotFound(error_message)
+
+    def get_component_info(self, component_name, spec):
+        error_message = ''
+        for storage_url in self.storage_urls:
+            api_client = StorageClient(storage_url, sources=self.sources)
+            try:
+                info = api_client.get_component_info(component_name=component_name, spec=spec)
+                return ComponentInfo(info, storage_url)
+            except (VersionNotFound, ComponentNotFound) as err:
+                error_message = str(err)
+
+        raise VersionNotFound(error_message)
+
+    def version_dependencies(self, version):
+        storage_url = self.storage_urls[0]  # We need any storage url
+        api_client = StorageClient(storage_url, sources=self.sources)
+        return api_client.version_dependencies(version)