+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This changelog is managed with commitizen tool, don't update it manually.
 - Add URL and status code to network error messages
 - Added --repository and --commit-sha parameters for packing and uploading component
 - Added aliases for '--default_namespace' as '--default-namespace' and '--registry_url' as '--registry-url'. The previous versions have been marked as deprecated.
+- Added CLI command `compote registry sync` to download components and synchronize a local mirror
 
 ### Fixed